backend.service: refactor parcel service interface to a new file
